+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `--repeat` to `pymbd.benchmark`
 - pkg-config support for ELSI
 
 ### Removed